data-aquisition/requirements.txt
python-dotenv
urllib3
shapely
